Refactoring of TaskList and Ui class to their respective packages
